+++ tests/src/cws/test_cws.py
@@ -82,3 +82,9 @@ class TestClass:
                    args=['-p', example_dir, '-m', 'example', '-s', 'tech_app', 'export'], obj={})
         assert pytest_wrapped_e.type == SystemExit
         assert pytest_wrapped_e.value.code == 0
+
+    def test_multi(self, example_dir):
+        with pytest.raises(SystemExit) as pytest_wrapped_e:
+            client(prog_name='cws', args=['-p', example_dir, 'test'], obj={})
+        assert pytest_wrapped_e.type == SystemExit
+        assert pytest_wrapped_e.value.code == 0